fix for product slug candidates and history
